Fix TRK func signatures

+++ include/TRK_MINNOW_DOLPHIN/Os/dolphin/targsupp.h
@@ -7,10 +7,10 @@
 extern "C" {
 #endif
 
-u8 TRKAccessFile(u8 command, u32 file_handle, u32 *length_ptr, u8 *buffer_ptr);
-u8 TRKOpenFile(u8 command, u32 file_handle, u32 flags, u8 *buffer_ptr);
-u8 TRKCloseFile(u8 command, u32 file_handle);
-u8 TRKPositionFile(u8 command, u32 file_handle, u32 *length_ptr, u8 *buffer_ptr);
+u32 TRKAccessFile(u8 command, u32 file_handle, u32 *length_ptr, u8 *buffer_ptr);
+u32 TRKOpenFile(u8 command, u32 file_handle, u32 flags, u8 *buffer_ptr);
+u32 TRKCloseFile(u8 command, u32 file_handle);
+u32 TRKPositionFile(u8 command, u32 file_handle, u32 *length_ptr, u8 *buffer_ptr);
 
 #ifdef __cplusplus
 }
